delete yarn project
