--- example/dev-proxy/package.json
@@ -1,20 +1,0 @@
-{
-  "name": "dev-proxy",
-  "version": "1.0.0",
-  "main": "index.js",
-  "license": "MIT",
-  "devDependencies": {
-    "@types/express": "^4.17.21",
-    "@types/node": "^20.12.12",
-    "typescript": "^5.4.5"
-  },
-  "dependencies": {
-    "axios": "^1.6.8",
-    "dotenv": "^16.4.5",
-    "express": "^4.19.2"
-  },
-  "scripts": {
-    "build": "tsc",
-    "start": "node dist/app"
-  }
-}
