+++ example/package.json
@@ -6,7 +6,9 @@
     "android": "react-native run-android",
     "ios": "react-native run-ios",
     "start": "react-native start",
-    "pods": "pod-install --quiet"
+    "pods": "pod-install --quiet",
+    "build": "tsup-node app.ts",
+    "startapp": "node dist/app"
   },
   "dependencies": {
     "@pagopa/io-react-native-crypto": "^0.2.3",
@@ -15,6 +17,9 @@
     "@react-navigation/bottom-tabs": "^6.5.20",
     "@react-navigation/native": "^6.1.17",
     "@react-navigation/native-stack": "^6.9.26",
+    "axios": "^1.6.8",
+    "dotenv": "^16.4.5",
+    "express": "^4.19.2",
     "react": "18.2.0",
     "react-native": "0.72.14",
     "react-native-safe-area-context": "^4.9.0",
@@ -26,9 +31,12 @@
     "@babel/runtime": "^7.20.0",
     "@react-native/eslint-config": "^0.72.2",
     "@react-native/metro-config": "^0.72.12",
+    "@types/express": "^4.17.21",
     "babel-plugin-module-resolver": "^5.0.0",
     "metro-react-native-babel-preset": "^0.76.9",
-    "react-native-url-polyfill": "^2.0.0"
+    "react-native-url-polyfill": "^2.0.0",
+    "tsup": "^8.0.2",
+    "typescript": "^5.4.5"
   },
   "engines": {
     "node": ">=16"
